Find and draw Contours on ROI of an image
Find and draw Contours on the ROI of an image:
We divide our image into the different sectors according to our needs.  FInd ROI and drow counter is very important for any detecting, tracking, or finding items in OpenCV programs. Take an idea from this simple code and apply it according to your need.